fastapi/requirements.txt
fastapi==0.68.0
python-multipart==0.0.5 
aiofiles==0.4.0
numpy
gunicorn
uvicorn
Pillow
scipy
facenet-pytorch
ultralytics